+++ app/code/ThirdParty/BlogArticle/Api/Data/CategoryInterface.php
@@ -6,6 +6,7 @@ interface CategoryInterface
     public const CATEGORY_ID = 'category_id';
     public const NAME = 'name';
     public const URL_KEY = 'url_key';
+    public const DESCRIPTION = 'description';
     public const IS_ACTIVE = 'is_active';
     public const CREATION_TIME = 'creation_time';
     public const UPDATE_TIME = 'update_time';
@@ -42,6 +43,17 @@ interface CategoryInterface
      * @return $this
      */
     public function setUrlKey($urlKey);
+
+    /**
+     * @return string|null
+     */
+    public function getDescription();
+
+    /**
+     * @param string|null $description
+     * @return $this
+     */
+    public function setDescription($description);
 
     /**
      * @return int|null