+++ app/code/ThirdParty/BlogArticle/Model/Resolver/Category.php
@@ -50,6 +50,7 @@ class Category implements ResolverInterface
             'category_id' => $category->getCategoryId(),
             'name' => $category->getName(),
             'url_key' => $category->getUrlKey(),
+            'description' => $category->getDescription(),
             'is_active' => $category->getIsActive(),
             'creation_time' => $category->getCreationTime(),
             'update_time' => $category->getUpdateTime(),